Move monitor config
